--- src/main.browerRouter.js
@@ -1,7 +1,0 @@
-//이곳에 구현 후 테스트 예정
-
-import createRouter from "./routes";
-
-const browserRouter = createRouter("browser");
-
-browserRouter.init();
