--- src/main.hashRouter.js
@@ -1,5 +1,0 @@
-import createRouter from "./routes";
-
-const hashRouter = createRouter("hash");
-
-hashRouter.init();
